Updating the extension without webpack base build system added packaging methods
