--- runtime/runtime-ui/.prettierignore
@@ -1,4 +1,0 @@
-.vscode-test
-out
-syntaxes/
-dist
