--- runtime/runtime-ui/.prettierrc.json
@@ -1,7 +1,0 @@
-{
-  "arrowParens": "always",
-  "printWidth": 130,
-  "singleQuote": true,
-  "trailingComma": "es5",
-  "endOfLine": "auto"
-}
