+++ runtime/runtime-ui/.vscodeignore
@@ -1,25 +1,10 @@
 .vscode/**
-typings/**
-out/**
-test/**
-src/**
-images/**
-node_modules/**
-scripts/**
-build/**
-Jenkinsfile
-prettier*
-**/*.map
-.gitignore
-.github/**
-tsconfig.json
-vsc-extension-quickstart.md
-undefined/**
-CONTRIBUTING.md
 .vscode-test/**
-.vscode-test-web/**
-**/**.vsix
-**/**.tar.gz
-!node_modules/prettier/index.js
-!node_modules/prettier/third-party.js
-!node_modules/prettier/parser-yaml.js
+src/**
+.gitignore
+.yarnrc
+vsc-extension-quickstart.md
+**/tsconfig.json
+**/.eslintrc.json
+**/*.map
+**/*.ts
