+++ runtime/runtime-ui/LICENSE.md
@@ -1,0 +1,9 @@
+Apache License
+Version 2.0, January 2004
+http://www.apache.org/licenses/
+
+APK Config Language Support
+
+This extension is licensed under the Apache License, Version 2.0. You may obtain a copy of the license at:
+
+[http://www.apache.org/licenses/LICENSE-2.0](http://www.apache.org/licenses/LICENSE-2.0)
