+++ runtime/runtime-ui/README.md
@@ -1,7 +1,43 @@
+# APK Config Language Support
 
-[![License](https://img.shields.io/badge/License-Apache_2.0-blue.svg)](https://github.com/wso2/apk/blob/main/LICENSE)
+> A Visual Studio Code extension for providing language support for APK configuration YAML files.
 
+## Features
 
-# APK Configuration support by WSO2
-Provides comprehensive YAML Language support to [Visual Studio Code](https://code.visualstudio.com/), via the [wso2-apk](https://github.com/wso2/apk).
+- Syntax highlighting for APK configuration YAML files.
+- Auto-completion for APK configuration YAML properties and values.
+- Validation and error checking for APK configuration YAML files.
+- Code snippets for common APK configuration patterns.
 
+## Requirements
+
+- Visual Studio Code version 1.0.0 or newer.
+
+## Installation
+
+1. Launch Visual Studio Code.
+2. Go to the Extensions view by clicking on the square icon in the sidebar or pressing `Ctrl+Shift+X`.
+3. Search for "APK Config Language Support".
+4. Click on the "Install" button for the extension published by "Your Name".
+5. The extension will be installed and activated automatically.
+
+## Usage
+
+1. Open a YAML file that contains APK configuration.
+2. The extension will automatically provide syntax highlighting and code completion for APK configuration properties and values.
+3. Errors and warnings will be displayed if there are any issues with the YAML structure or values.
+4. Use the provided code snippets to quickly insert common APK configuration patterns.
+
+## Contributing
+
+Contributions are welcome! Here's how you can get involved:
+
+1. Fork the repository.
+2. Create a new branch for your feature or bug fix.
+3. Make your changes and commit them.
+4. Push your changes to your fork.
+5. Submit a pull request describing your changes.
+
+## License
+
+This extension is licensed under the [Apache 2.0 License](LICENSE).
