--- runtime/runtime-ui/custom-settings.json
@@ -1,5 +1,0 @@
-{
-    "yaml.schemas": {
-        "https://json.schemastore.org/kustomization.json": "/*"
-    }
-}
