+++ runtime/runtime-ui/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "wso2-apk",
-  "displayName": "WSO2 APK",
-  "description": "WSO2 APK configuration Language Support.",
+  "name": "apk-config-language-support",
+  "displayName": "APK Config Language Support",
+  "description": "APK config language support ",
   "author": "WSO2",
-  "license": "MIT",
+  "license": "Apache 2.0",
   "version": "0.0.1",
   "publisher": "wso2",
   "bugs": "https://github.com/wso2/apk/issues",
@@ -14,8 +14,15 @@
   "icon": "icon/icon128.png",
   "engines": {
     "npm": ">=7.0.0",
-    "vscode": "^1.63.0"
+    "vscode": "^1.79.0"
   },
+  "keywords": [
+    "apk",
+    "yaml",
+    "autocompletion",
+    "validation",
+    "wso2"
+  ],
   "categories": [
     "Programming Languages",
     "Linters",
@@ -31,14 +38,7 @@
   "activationEvents": [
     "onLanguage:yaml"
   ],
-  "keywords": [
-    "apk",
-    "yaml",
-    "autocompletion",
-    "validation",
-    "wso2"
-  ],
-  "main": "./dist/extension",
+  "main": "./out/extension.js",
   "contributes": {
     "languages": [
       {
@@ -72,52 +72,14 @@
     "configuration": {},
     "configurationDefaults": {}
   },
-  "extensionDependencies": [],
   "scripts": {
-    "build": "yarn run clean && yarn run lint && yarn run vscode:prepublish",
-    "check-dependencies": "node ./scripts/check-dependencies.js",
-    "clean": "rimraf out && rimraf dist",
-    "compile": "webpack --mode none",
-    "format": "prettier --write .",
-    "lint": "eslint -c .eslintrc.js --ext .ts src test",
-    "vscode:prepublish": "webpack --mode production"
-  },
-  "devDependencies": {
-    "@types/chai": "^4.2.12",
-    "@types/fs-extra": "^9.0.6",
-    "@types/mocha": "^2.2.48",
-    "@types/node": "^12.12.6",
-    "@types/sinon": "^10.0.6",
-    "@types/sinon-chai": "^3.2.5",
-    "@types/vscode": "^1.63.0",
-    "@types/webpack": "^4.4.10",
-    "@typescript-eslint/eslint-plugin": "^4.16.1",
-    "@typescript-eslint/parser": "^4.16.1",
-    "@vscode/test-web": "0.0.11",
-    "buffer": "^6.0.3",
-    "chai": "^4.2.0",
-    "crypto-browserify": "^3.12.0",
-    "eslint": "^7.6.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "glob": "^7.1.6",
-    "mocha": "^9.1.2",
-    "path-browserify": "^1.0.1",
-    "prettier": "2.2.1",
-    "process": "^0.11.10",
-    "rimraf": "^3.0.2",
-    "sinon": "^12.0.1",
-    "sinon-chai": "^3.7.0",
-    "ts-loader": "^9.2.5",
-    "ts-node": "^3.3.0",
-    "typescript": "4.4.3",
-    "umd-compat-loader": "^2.1.2",
-    "url": "^0.11.0",
-    "util": "^0.12.5",
-    "vscode-extension-tester": "^5.3.0",
-    "vscode-test": "^1.4.0",
-    "webpack": "^5.76.1",
-    "webpack-cli": "^5.0.1"
+    "vscode:prepublish": "yarn run compile",
+    "compile": "tsc -p ./",
+    "watch": "tsc -watch -p ./",
+    "pretest": "yarn run compile && yarn run lint",
+    "lint": "eslint src --ext ts",
+    "test": "node ./out/test/runTest.js",
+    "build": "vsce package"
   },
   "dependencies": {
     "@redhat-developer/vscode-redhat-telemetry": "^0.6.0",
@@ -128,5 +90,45 @@
     "vscode-uri": "^2.0.3",
     "whatwg-fetch": "^3.6.2",
     "yaml-language-server": "next"
+  },
+  "devDependencies": {
+    "@types/chai": "^4.2.12",
+    "@types/fs-extra": "^9.0.6",
+    "@types/glob": "^8.1.0",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "20.2.5",
+    "@types/sinon": "^10.0.6",
+    "@types/sinon-chai": "^3.2.5",
+    "@types/vscode": "^1.79.0",
+    "@types/webpack": "^4.4.10",
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
+    "@vscode/test-electron": "^2.3.2",
+    "@vscode/test-web": "0.0.11",
+    "buffer": "^6.0.3",
+    "chai": "^4.2.0",
+    "crypto-browserify": "^3.12.0",
+    "eslint": "^8.41.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "glob": "^8.1.0",
+    "mocha": "^10.2.0",
+    "path-browserify": "^1.0.1",
+    "prettier": "2.2.1",
+    "process": "^0.11.10",
+    "rimraf": "^3.0.2",
+    "sinon": "^12.0.1",
+    "sinon-chai": "^3.7.0",
+    "ts-loader": "^9.2.5",
+    "ts-node": "^3.3.0",
+    "typescript": "^5.1.3",
+    "umd-compat-loader": "^2.1.2",
+    "url": "^0.11.0",
+    "util": "^0.12.5",
+    "vsce": "^2.15.0",
+    "vscode-extension-tester": "^5.3.0",
+    "vscode-test": "^1.4.0",
+    "webpack": "^5.76.1",
+    "webpack-cli": "^5.0.1"
   }
 }
