--- runtime/runtime-ui/scripts/e2e.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-node "$(pwd)/out/test/testRunner"
